requirements.txt
pdfPlumber
nltk
faiss-cpu
sentence-transformers
streamlit